move the provisioning error handling to provisioning json
